// amplify/backend/function/jobUpdateOwner/src/API.ts
export enum EmployeeType {
  OTHER = "OTHER",
  NORMAL = "NORMAL",
  MANAGER = "MANAGER",
  SUPER_MANAGER = "SUPER_MANAGER",
}


export enum BooleanType {
  TRUE = "TRUE",
  FALSE = "FALSE",
}


export type Employee = {
  __typename: "Employee",
  username: string,
  companyID: string,
  no: string,
  groupID: string,
  superiorUsername: string,
  firstName: string,
  lastName: string,
  grade: string,
  email: string,
  sub?: string | null,
  manager: EmployeeType,
  isCompanyAdmin?: boolean | null,
  isDeleted: BooleanType,
  createdAt: string,
  updatedAt: string,
  company?: Company | null,
  superior?: Employee | null,
  group?: Group | null,
};

export type Company = {
  __typename: "Company",
  id: string,
  name: string,
  startMonth: number,
  shortName?: string | null,
  url?: string | null,
  isContractEvaluation?: boolean | null,
  isContractReport?: boolean | null,
  createdAt: string,
  updatedAt: string,
};

export type Group = {
  __typename: "Group",
  id: string,
  companyID: string,
  no: string,
  name: string,
  createdAt: string,
  updatedAt: string,
};